backend/requirements.txt
flask
flask-cors
flask-login
flask_sqlalchemy
flask_wtf
passlib.hash
wtforms
wtforms.validators